Re-implementation with JS and routing
I re-did the project with JS in the settings instead of TS. I also completed the routing with all the imports where necessary.

The component has been copied from the general predefined components folder to the domains/groceries/components.

Edit and Create are still empty, but the table and the navigation work.